Add alerting system with Slack and PagerDuty integration
Introduces a configurable alerting subsystem supporting Slack and PagerDuty sinks, including deduplication and queueing. Alerts are emitted on policy enforcement events and failures in both HTTP and gRPC APIs, as well as in the policy enforcer. Configuration options are added to the example config file, and alerting can be tested via a new CLI command. Includes unit tests for dispatcher and sinks.